+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
+    * HOTFIX      #2401 [MediaBundle]         Fixed slow media queries
     * HOTFIX      #2401 [MediaBundle]         Fixed search in media bundle
     * HOTFIX      #2381 [ContentBundle]       Fixed auto-name subscriber to rename at the very end of persist
     * HOTFIX      #2388 [Rest]                Fixed bug when applying same sortfield multiple times